Redo events page
The DSST events have been merged with the events page.

Placeholders for the Machine Learning in Brain Imaging Series have been
added to the events page.

The events page is now in reverse chronological order, grouped by month.

The DSST and MLT pages are no longer in a two column layout.